Triage Error Prone ReferenceEquality and numeric boxing/narrowing warnings (#3601)
ReferenceEquality (4): all intentional, suppressed with justification -
memoized-supplier identity checks in a test, JNA's own singleton
TypeMapper comparison idiom, and a Swing component-identity check.

BoxingComparator (8): Comparator.comparing(...) boxing an int/long key
extractor; switched to comparingInt/comparingLong.

NarrowCalculation (9): products of small, OS-bounded int quantities
(fd counts, process counts, tiny struct-offset constants) that can't
realistically overflow; suppressed with the bound documented.

LongDoubleConversion (5) and IntLiteralCast (3): missing explicit casts
on long-to-double widening and int-literal-to-long casts; made explicit
or replaced with a typed literal, no behavior change.

UnnecessaryLongToIntConversion (4): removed pointless (int) casts on
long values immediately passed to a long-typed parameter.

IntLongMath (1): MacOSProcessJNA computed pti_faults - pti_pageins in
int arithmetic before widening to the long minorFaults field. Its FFM
twin already avoids this by subtracting from the already-long
majorFaults field instead of the raw int; changed JNA to match rather
than suppress, since it removes the pattern entirely.

+++ oshi-core-ffm/src/main/java/oshi/software/os/windows/WindowsOSProcessFFM.java
@@ -336,7 +336,7 @@ public class WindowsOSProcessFFM extends WindowsOSProcess {
                     if (envSize > 0 && envSize < Integer.MAX_VALUE) {
                         MemorySegment envAddress = upp.get(ADDRESS, UPP_ENVIRONMENT_OFFSET);
                         if (envAddress.address() != 0) {
-                            MemorySegment envBuffer = arena.allocate((int) envSize);
+                            MemorySegment envBuffer = arena.allocate(envSize);
                             if (Kernel32FFM.ReadProcessMemory(h.get(), envAddress, envBuffer, envSize, bytesRead)
                                     && bytesRead.get(JAVA_LONG, 0) > 0) {
                                 char[] env = new char[(int) (envSize / 2)];
